+++ changelog.md
@@ -15,6 +15,7 @@
 * Tweak - Small changes to allow LSX Customizer to change WooCommerce cart menu item position
 * Tweak - Small changes to allow LSX Customizer to change WooCommerce cart menu item style
 * Tweak - Small changes to allow LSX Customizer to change WooCommerce My Account menu item
+* Tweak - Add WooCommerce Wishlist compatibility
 
 ### 2.0.0
 * New totally awesome version! The visual was fully redesigned